requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
httpx>=0.27.0
Pillow>=10.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
